Move codec to a module
